refactor: Use correct namespaces in Menu
